+++ DependencyInjection/DeveloptechAgilityExtension.php
@@ -23,6 +23,7 @@ class DeveloptechAgilityExtension extends Extension
         $config = $this->processConfiguration($configuration, $configs);
 
         $container->setParameter('developtech_agility.model.project.class', $config['model']['class']['project']);
+        $container->setParameter('developtech_agility.model.feature.class', $config['model']['class']['feature']);
 
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('services.yml');