+++ Tests/Mock/Feature.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Developtech\AgilityBundle\Tests\Mock;
+
+use Developtech\AgilityBundle\Model\FeatureModel;
+
+class Feature extends FeatureModel {
+
+}