+++ CHANGELOG.md
@@ -4,4 +4,6 @@
 * Projects list service
 * Get project service
 * Project creation service
+* Feature model
+* Project features list service
 * Setup documentation